refactor: separate business logic in Carousel using useCarousel hook
