+++ src/components/Carousel/Carousel.tsx
@@ -1,9 +1,9 @@
-import { useEffect, useRef, useState } from 'react';
 import classNames from 'classnames';
 
 import TravelProduct, { TravelProductData } from '../TravelProduct/TravelProduct';
 
 import styles from './Carousel.module.css';
+import useCarousel from './useCarousel';
 
 export type ItemSize = 'small' | 'middle' | 'large';
 type CarouselType<T> = {
@@ -11,44 +11,21 @@ type CarouselType<T> = {
   itemSize: ItemSize;
 };
 
-const ITEM_SIZE = {
-  small: 210,
-  middle: 238,
-  large: 267,
-};
-const GAP_SIZE = 24;
+// const ITEM_SIZE = {
+//   small: 210,
+//   middle: 238,
+//   large: 267,
+// };
+// const GAP_SIZE = 24;
 
 const Carousel = <T extends TravelProductData>({ items, itemSize }: CarouselType<T>) => {
-  const carouselListRef = useRef<HTMLUListElement | null>(null);
-  const [slideCount, setSlideCount] = useState(0);
-
-  const isStartPoint = slideCount <= 0;
-  const isEndPoint =
-    carouselListRef.current === null
-      ? false
-      : (ITEM_SIZE[itemSize] + GAP_SIZE) * slideCount + carouselListRef.current.offsetWidth >
-        carouselListRef.current.scrollWidth;
-
-  const handleClickPrevSlideButton = () => {
-    if (isStartPoint) return;
-
-    setSlideCount((prev) => prev - 1);
-  };
-
-  const handleClickNextSlideButton = () => {
-    if (isEndPoint) return;
-
-    setSlideCount((prev) => prev + 1);
-  };
-
-  useEffect(() => {
-    if (!carouselListRef.current) return;
-
-    carouselListRef.current.scrollTo({
-      left: (ITEM_SIZE[itemSize] + GAP_SIZE) * slideCount,
-      behavior: 'smooth',
-    });
-  }, [slideCount]);
+  const {
+    isStartPoint,
+    isEndPoint,
+    carouselListRef,
+    handleClickPrevSlideButton,
+    handleClickNextSlideButton,
+  } = useCarousel(itemSize);
 
   return (
     <div className={styles.box}>
